+++ app/src/main/java/com/peryite/journeyd3/models/Conquest.java
@@ -4,11 +4,9 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-//@ToString
-//@EqualsAndHashCode
 @NoArgsConstructor
 public class Conquest {
-    private int id;
+    private long id;
     private String name;
     private String description;
     private boolean done;
